Make a module for the concrete P instance.
