+++ matching/src/Pattern/Type.hs
@@ -1,8 +1,3 @@
-{-# LANGUAGE DeriveFunctor     #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE TemplateHaskell   #-}
-{-# LANGUAGE OverloadedStrings   #-}
-
 module Pattern.Type
   ( Column (..)
   , Metadata (..)
@@ -10,14 +5,10 @@ module Pattern.Type
   , Action (..)
   , ClauseMatrix (..)
   , Clause (..)
-  , P (..)
-  , Pattern
-  , BoundPattern
   , PatternMatrix (..)
   , VariableBinding (..)
   , Ignoring (..)
   , Fringe
-  , Index
   , Constructor (..)
   , Symbol (..)
   , Literal (..)
@@ -25,21 +16,17 @@ module Pattern.Type
   , getConstructorPattern
   ) where
 
-import Data.Deriving
-       ( deriveEq2, deriveOrd2, deriveShow2 )
 import Data.Functor.Classes
-       ( Eq1 (..), Ord1 (..), Show1 (..), liftCompare2, liftEq2,
-       liftShowsPrec2 )
+       ( Show1 (..) )
 import Data.Functor.Foldable
-       ( Fix (..), cata )
+       ( Fix (..) )
 import Data.Text
        ( Text, pack )
 import Kore.AST.Common
        ( Sort (..), SymbolOrAlias (..) )
 import Kore.AST.MetaOrObject
        ( Object (..) )
-import           Kore.Unparser
-                 ( unparseToString )
+
 import qualified Data.Yaml.Builder as Y
 
 data Column p bp = Column
@@ -87,42 +74,7 @@ serializeOccurrence (Inj o) = "-1" : serializeOccurrence o
 instance Y.ToYaml Occurrence where
   toYaml o = Y.toYaml $ serializeOccurrence o
 
-instance Y.ToYaml a => Y.ToYaml (BoundPattern a) where
-  toYaml Wildcard = error "Unsupported map/set pattern"
-  toYaml (Variable (Just o) h) = Y.mapping
-    ["hook" Y..= Y.toYaml (pack h)
-    , "occurrence" Y..= Y.toYaml o
-    ]
-  toYaml (Variable Nothing _) = error "Unsupported map/set pattern"
-  toYaml (As _ _ p) = Y.toYaml p
-  toYaml (MapPattern _ _ _ _ o) = Y.toYaml o
-  toYaml (SetPattern _ _ _ o) = Y.toYaml o
-  toYaml (ListPattern _ _ _ _ o) = Y.toYaml o
-  toYaml (Pattern (Right (Literal s)) (Just h) []) = Y.mapping
-    ["hook" Y..= Y.toYaml (pack h)
-    , "literal" Y..= Y.toYaml (pack s)
-    ]
-  toYaml (Pattern (Left (Symbol s)) Nothing ps) = Y.mapping
-    ["constructor" Y..= Y.toYaml (pack $ unparseToString s)
-    , "args" Y..= Y.array (map Y.toYaml ps)
-    ]
-  toYaml Pattern{} = error "Unsupported map/set pattern"
-
-
-instance Y.ToYaml (Fix BoundPattern) where
-  toYaml = cata Y.toYaml
-
-
 type Fringe = [(Occurrence, Bool)] -- occurrence and whether to match the exact sort
-
-instance Show1 Pattern where
-  liftShowsPrec = liftShowsPrec2 showsPrec showList
-instance Show1 BoundPattern where
-  liftShowsPrec = liftShowsPrec2 showsPrec showList
-instance Eq1 BoundPattern where
-  liftEq = liftEq2 (==)
-instance Ord1 BoundPattern where
-  liftCompare = liftCompare2 compare
 
 newtype Symbol = Symbol (SymbolOrAlias Object)
                  deriving (Eq, Show, Ord)
@@ -161,39 +113,6 @@ getConstructorPattern (SymbolConstructor s)  = Left s
 getConstructorPattern (LiteralConstructor l) = Right l
 getConstructorPattern _ = error "Pattern constructor must be Symbol or Literal"
 
-type Index       = Int
-data P var a   = Pattern (Either Symbol Literal) (Maybe String) ![a]
-                 | ListPattern
-                   { getHead :: ![a] -- match elements at front of list
-                   , getFrame :: !(Maybe a) -- match remainder of list
-                   , getTail :: ![a] -- match elements at back of list
-                   , element :: SymbolOrAlias Object -- ListItem symbol
-                   , original :: a
-                   }
-                 | MapPattern
-                   { getKeys :: ![a]
-                   , getValues :: ![a]
-                   , getFrame :: !(Maybe a)
-                   , element :: SymbolOrAlias Object
-                   , original :: !a
-                   }
-                 | SetPattern
-                   { getElements :: ![a]
-                   , getFrame :: !(Maybe a)
-                   , element :: SymbolOrAlias Object
-                   , original :: !a
-                   }
-                 | As var String a
-                 | Wildcard
-                 | Variable var String
-                 deriving (Show, Eq, Functor)
-
-type Pattern = P String
-type BoundPattern = P (Maybe Occurrence)
-
-$(deriveEq2 ''P)
-$(deriveShow2 ''P)
-$(deriveOrd2 ''P)
 
 newtype PatternMatrix p bp = PatternMatrix [Column p bp]
                              deriving (Show)
